update roles, filters, and profiles
