+++ client/src/utils/token.js
@@ -1,0 +1,12 @@
+export const getToken = () => {
+  return localStorage.getItem("token");
+};
+
+export const setToken = (token) => {
+  localStorage.setItem("token", token);
+};
+
+export const clearToken = () => {
+  localStorage.clearItem("token");
+};
+
